Remove warnings

+++ LIBRARY/src/bid32_to_int32.c
@@ -218,22 +218,22 @@ bid32_to_int32_rnint (BID_UINT32 x _EXC_FLAGS_PARAM _EXC_MASKS_PARAM
     }	// else MP in [ODD, EVEN]
       }
       if (x_sign)
-    res = -Cstar;
+    res = (int) (-Cstar);
       else
-    res = Cstar;
+    res = (int) Cstar;
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +/-C (exact)
       if (x_sign)
-    res = -C1;
+    res = (int) (-C1);
       else
-    res = C1;
+    res = (int) C1;
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +/-C * 10^exp (exact)
       if (x_sign)
-    res = -C1 * bid_ten2k64[exp];
+    res = (int) (-C1 * bid_ten2k64[exp]);
       else
-    res = C1 * bid_ten2k64[exp];
+    res = (int) (C1 * bid_ten2k64[exp]);
     }
   }
   BID_RETURN (res);
@@ -468,22 +468,22 @@ bid32_to_int32_xrnint (BID_UINT32 x _EXC_FLAGS_PARAM _EXC_MASKS_PARAM
     }	// else MP in [ODD, EVEN]
       }
       if (x_sign)
-    res = -Cstar;
+    res = (int) (-Cstar);
       else
-    res = Cstar;
+    res = (int) Cstar;
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +/-C (exact)
       if (x_sign)
-    res = -C1;
+    res = (int) (-C1);
       else
-    res = C1;
+    res = (int) C1;
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +/-C * 10^exp (exact)
       if (x_sign)
-    res = -C1 * bid_ten2k64[exp];
+    res = (int) (-C1 * bid_ten2k64[exp]);
       else
-    res = C1 * bid_ten2k64[exp];
+    res = (int) (C1 * bid_ten2k64[exp]);
     }
   }
   BID_RETURN (res);
@@ -670,22 +670,22 @@ bid32_to_int32_floor (BID_UINT32 x _EXC_FLAGS_PARAM _EXC_MASKS_PARAM
       }
 
       if (x_sign)
-    res = -Cstar;
+    res = (int) (-Cstar);
       else
-    res = Cstar;
+    res = (int) Cstar;
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +/-C (exact)
       if (x_sign)
-    res = -C1;
+    res = (int) (-C1);
       else
-    res = C1;
+    res = (int) C1;
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +/-C * 10^exp (exact)
       if (x_sign)
-    res = -C1 * bid_ten2k64[exp];
+    res = (int) (-C1 * bid_ten2k64[exp]);
       else
-    res = C1 * bid_ten2k64[exp];
+    res = (int) (C1 * bid_ten2k64[exp]);
     }
   }
   BID_RETURN (res);
@@ -878,22 +878,22 @@ bid32_to_int32_xfloor (BID_UINT32 x _EXC_FLAGS_PARAM _EXC_MASKS_PARAM
       }
 
       if (x_sign)
-    res = -Cstar;
+    res = (int) (-Cstar);
       else
-    res = Cstar;
+    res = (int) Cstar;
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +/-C (exact)
       if (x_sign)
-    res = -C1;
+    res = (int) (-C1);
       else
-    res = C1;
+    res = (int) C1;
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +/-C * 10^exp (exact)
       if (x_sign)
-    res = -C1 * bid_ten2k64[exp];
+    res = (int) (-C1 * bid_ten2k64[exp]);
       else
-    res = C1 * bid_ten2k64[exp];
+    res = (int) (C1 * bid_ten2k64[exp]);
     }
   }
   BID_RETURN (res);
@@ -1080,22 +1080,22 @@ bid32_to_int32_ceil (BID_UINT32 x _EXC_FLAGS_PARAM _EXC_MASKS_PARAM
       }
 
       if (x_sign)
-    res = -Cstar;
+    res = (int) (-Cstar);
       else
-    res = Cstar;
+    res = (int) Cstar;
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +/-C (exact)
       if (x_sign)
-    res = -C1;
+    res = (int) (-C1);
       else
-    res = C1;
+    res = (int) C1;
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +/-C * 10^exp (exact)
       if (x_sign)
-    res = -C1 * bid_ten2k64[exp];
+    res = (int) (-C1 * bid_ten2k64[exp]);
       else
-    res = C1 * bid_ten2k64[exp];
+    res = (int) (C1 * bid_ten2k64[exp]);
     }
   }
   BID_RETURN (res);
@@ -1288,22 +1288,22 @@ bid32_to_int32_xceil (BID_UINT32 x _EXC_FLAGS_PARAM _EXC_MASKS_PARAM
       }
 
       if (x_sign)
-    res = -Cstar;
+    res = (int) (-Cstar);
       else
-    res = Cstar;
+    res = (int) Cstar;
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +/-C (exact)
       if (x_sign)
-    res = -C1;
+    res = (int) (-C1);
       else
-    res = C1;
+    res = (int) C1;
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +/-C * 10^exp (exact)
       if (x_sign)
-    res = -C1 * bid_ten2k64[exp];
+    res = (int) (-C1 * bid_ten2k64[exp]);
       else
-    res = C1 * bid_ten2k64[exp];
+    res = (int) (C1 * bid_ten2k64[exp]);
     }
   }
   BID_RETURN (res);
@@ -1461,22 +1461,22 @@ bid32_to_int32_int (BID_UINT32 x _EXC_FLAGS_PARAM _EXC_MASKS_PARAM
       shift = bid_shiftright128[ind - 1];	// 0 <= shift <= 39
       Cstar = Cstar >> shift;
       if (x_sign)
-    res = -Cstar;
+    res = (int) (-Cstar);
       else
-    res = Cstar;
+    res = (int) Cstar;
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +/-C (exact)
       if (x_sign)
-    res = -C1;
+    res = (int) (-C1);
       else
-    res = C1;
+    res = (int) C1;
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +/-C * 10^exp (exact)
       if (x_sign)
-    res = -C1 * bid_ten2k64[exp];
+    res = (int) (-C1 * bid_ten2k64[exp]);
       else
-    res = C1 * bid_ten2k64[exp];
+    res = (int) (C1 * bid_ten2k64[exp]);
     }
   }
   BID_RETURN (res);
@@ -1660,22 +1660,22 @@ bid32_to_int32_xint (BID_UINT32 x _EXC_FLAGS_PARAM _EXC_MASKS_PARAM
       }
 
       if (x_sign)
-    res = -Cstar;
+    res = (int) (-Cstar);
       else
-    res = Cstar;
+    res = (int) Cstar;
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +/-C (exact)
       if (x_sign)
-    res = -C1;
+    res = (int) (-C1);
       else
-    res = C1;
+    res = (int) C1;
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +/-C * 10^exp (exact)
       if (x_sign)
-    res = -C1 * bid_ten2k64[exp];
+    res = (int) (-C1 * bid_ten2k64[exp]);
       else
-    res = C1 * bid_ten2k64[exp];
+    res = (int) (C1 * bid_ten2k64[exp]);
     }
   }
   BID_RETURN (res);
@@ -1849,22 +1849,22 @@ bid32_to_int32_rninta (BID_UINT32 x _EXC_FLAGS_PARAM _EXC_MASKS_PARAM
 
       // if the result was a midpoint it was rounded away from zero
       if (x_sign)
-    res = -Cstar;
+    res = (int) (-Cstar);
       else
-    res = Cstar;
+    res = (int) Cstar;
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +/-C (exact)
       if (x_sign)
-    res = -C1;
+    res = (int) (-C1);
       else
-    res = C1;
+    res = (int) C1;
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +/-C * 10^exp (exact)
       if (x_sign)
-    res = -C1 * bid_ten2k64[exp];
+    res = (int) (-C1 * bid_ten2k64[exp]);
       else
-    res = C1 * bid_ten2k64[exp];
+    res = (int) (C1 * bid_ten2k64[exp]);
     }
   }
   BID_RETURN (res);
@@ -2081,22 +2081,22 @@ bid32_to_int32_xrninta (BID_UINT32 x _EXC_FLAGS_PARAM _EXC_MASKS_PARAM
 
       // if the result was a midpoint it was rounded away from zero
       if (x_sign)
-    res = -Cstar;
+    res = (int) (-Cstar);
       else
-    res = Cstar;
+    res = (int) Cstar;
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +/-C (exact)
       if (x_sign)
-    res = -C1;
+    res = (int) (-C1);
       else
-    res = C1;
+    res = (int) C1;
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +/-C * 10^exp (exact)
       if (x_sign)
-    res = -C1 * bid_ten2k64[exp];
+    res = (int) (-C1 * bid_ten2k64[exp]);
       else
-    res = C1 * bid_ten2k64[exp];
+    res = (int) (C1 * bid_ten2k64[exp]);
     }
   }
   BID_RETURN (res);
